Rename DisconnectableDirective to AsyncDirective
