+++ .prettierignore
@@ -34,7 +34,7 @@ packages/lit/test/
 packages/lit/decorators.*
 packages/lit/directive-helpers.*
 packages/lit/directive.*
-packages/lit/disconnectable-directive.*
+packages/lit/async-directive.*
 packages/lit/html.*
 packages/lit/hydrate-support.*
 packages/lit/hydrate.*
@@ -68,7 +68,7 @@ packages/lit-html/hydrate.*
 packages/lit-html/lit-html.*
 packages/lit-html/directive.*
 packages/lit-html/directive-helpers.*
-packages/lit-html/disconnectable-directive.*
+packages/lit-html/async-directive.*
 packages/lit-html/polyfill-support.*
 packages/lit-html/private-ssr-support.*
 packages/lit-html/static.*
