+++ packages/lit-html/.gitignore
@@ -6,7 +6,7 @@
 /lit-html.*
 /directive.*
 /directive-helpers.*
-/disconnectable-directive.*
+/async-directive.*
 /polyfill-support.*
 /private-ssr-support.*
 /static.*
