+++ packages/lit/package.json
@@ -31,8 +31,8 @@
     "./directives/": {
       "default": "./directives/"
     },
-    "./disconnectable-directive.js": {
-      "default": "./disconnectable-directive.js"
+    "./async-directive.js": {
+      "default": "./async-directive.js"
     },
     "./html.js": {
       "default": "./html.js"
@@ -55,7 +55,7 @@
     "build:watch": "rollup -c --watch",
     "build:ts": "tsc",
     "build:ts:watch": "tsc --watch",
-    "clean": "rm -rf decorators.{d.ts.map,d.ts,js.map,js} decorators/ directive-helpers.{d.ts.map,d.ts,js.map,js} directive.{d.ts.map,d.ts,js.map,js} directives/ disconnectable-directive.{d.ts.map,d.ts,js.map,js} html.{d.ts.map,d.ts,js.map,js} hydrate-support.{d.ts.map,d.ts,js.map,js} hydrate.{d.ts.map,d.ts,js.map,js} index.{d.ts.map,d.ts,js.map,js} lit.min.{js,js.map,d.ts} polyfill-support.{d.ts.map,d.ts,js.map,js} static-html.{d.ts.map,d.ts,js.map,js}",
+    "clean": "rm -rf decorators.{d.ts.map,d.ts,js.map,js} decorators/ directive-helpers.{d.ts.map,d.ts,js.map,js} directive.{d.ts.map,d.ts,js.map,js} directives/ async-directive.{d.ts.map,d.ts,js.map,js} html.{d.ts.map,d.ts,js.map,js} hydrate-support.{d.ts.map,d.ts,js.map,js} hydrate.{d.ts.map,d.ts,js.map,js} index.{d.ts.map,d.ts,js.map,js} lit.min.{js,js.map,d.ts} polyfill-support.{d.ts.map,d.ts,js.map,js} static-html.{d.ts.map,d.ts,js.map,js}",
     "publish-dev": "VERSION=${npm_package_version%-*}-dev.`git rev-parse --short HEAD` && npm version --no-git-tag-version $VERSION && npm publish --tag dev",
     "regen-package-lock": "rm -rf node_modules package-lock.json; npm install",
     "test": "MODE=prod cd ../tests && npx wtr '../lit/test/**/*_test.(js|html)'"
@@ -66,7 +66,7 @@
     "/directive-helpers.{d.ts.map,d.ts,js.map,js}",
     "/directive.{d.ts.map,d.ts,js.map,js}",
     "/directives/",
-    "/disconnectable-directive.{d.ts.map,d.ts,js.map,js}",
+    "/async-directive.{d.ts.map,d.ts,js.map,js}",
     "/html.{d.ts.map,d.ts,js.map,js}",
     "/hydrate-support.{d.ts.map,d.ts,js.map,js}",
     "/hydrate.{d.ts.map,d.ts,js.map,js}",
